# swagger/po-common/po-api.yml
swagger: '2.0'
info:
  version: 0.0.1
  title: Covid19 Citizen Portal
  description: Covid19 Citizen Portal
host: covid19.int.xcp
basePath: /api/v1
schemes:
  - https
  - http
tags:
  - name: person-controller
    description: Controller for Person data
  - name: health-check-controller
    description: Controller for Person's health check
  - name: treatment-controller
    description: Controller for treatment
  - name: authorization-controller
    description: Controller for Authorization
securityDefinitions:
  apiKey:
    type: apiKey
    name: Authorization
    in: header
paths:
  '/persons/{personUid}':
    get:
      tags:
        - person-controller
      summary: GET actual person and health status
      description: ''
      produces:
        - application/json
      parameters:
        - name: personUid
          in: path
          description: Unique Person's ID (person_uid.person)
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/PersonResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - apiKey: []
    put:
      tags:
        - person-controller
      summary: Updates actual person's and health status
      description: ''
      produces:
        - application/json
      parameters:
        - name: personUid
          in: path
          description: Unique Person's ID (person_uid.person)
          required: true
          type: string
          format: uuid
        - in: body
          name: personDto
          description: Person's data
          required: true
          schema:
            $ref: '#/definitions/PersonRequest'
      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/PersonResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - apiKey: []
  '/persons/{personUid}/health-check/symptoms':
    put:
      tags:
        - health-check-controller
      summary: Fills actual health check form
      description: ''
      produces:
        - application/json
      parameters:
        - name: personUid
          in: path
          description: Unique Person's ID (person_uid.person)
          required: true
          type: string
          format: uuid
        - in: body
          name: simptomsDto
          description: Health check's data - simptoms
          required: true
          schema:
            $ref: '#/definitions/SimtompsRequest'
      responses:
        '200':
          description: Successful operation
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - apiKey: []
  '/persons/{personUid}/health-check/exposure':
    put:
      tags:
        - health-check-controller
      summary: Fills exposure form
      description: ''
      produces:
        - application/json
      parameters:
        - name: personUid
          in: path
          description: Unique Person's ID (person_uid.person)
          required: true
          type: string
          format: uuid
        - in: body
          name: exposureDto
          description: Health check's data - exposure
          required: true
          schema:
            $ref: '#/definitions/ExposureRequest'
      responses:
        '200':
          description: Successful operation
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - apiKey: []
  '/persons/{personUid}/health-check/testing-place':
    put:
      tags:
        - health-check-controller
      summary: Fills testing place form
      description: ''
      produces:
        - application/json
      parameters:
        - name: personUid
          in: path
          description: Unique Person's ID (person_uid.person)
          required: true
          type: string
          format: uuid
        - in: body
          name: testingPlaceDto
          description: Health check's data - testing place
          required: true
          schema:
            $ref: '#/definitions/TestingPlaceRequest'
      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/TestingPlaceInstuctionsDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - apiKey: []
  '/persons/{personUid}/treatment':
    put:
      tags:
        - treatment-controller
      summary: Fills actual data about treatment for covid19 positive patients
      description: ''
      produces:
        - application/json
      parameters:
        - name: personUid
          in: path
          description: Unique Person's ID (person_uid.person)
          required: true
          type: string
          format: uuid
        - in: body
          name: TreatmentDto
          description: Treatment's data
          required: true
          schema:
            $ref: '#/definitions/TreatmentRequest'
      responses:
        '200':
          description: Successful operation
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - apiKey: []
  /authorizations/send-code:
    post:
      tags:
        - authorization-controller
      summary: Sent SMS with auth code
      operationId: sendCodeUsingPOST
      produces:
        - application/json
      parameters:
        - name: personUid
          in: query
          description: Uid of person
          required: false
          type: string
          format: uuid
          allowEmptyValue: false
      responses:
        '200':
          description: OK
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - apiKey: []
      deprecated: false
  /authorizations/verify-code:
    post:
      tags:
        - authorization-controller
      summary: Verify code from SMS
      operationId: verifyCodeUsingPOST
      produces:
        - application/json
      parameters:
        - name: personUid
          in: query
          description: Uid of person
          required: false
          type: string
          format: uuid
          allowEmptyValue: false
        - name: smsCode
          in: query
          description: Code received in SMS
          required: false
          type: string
          allowEmptyValue: false
      responses:
        '200':
          description: 'SMS code is OK, return password for login'
          schema:
            $ref: '#/definitions/VerifyCodeResponseDto'
        '201':
          description: Created
        '400':
          description: 'Some problem with SMS code, see description'
          schema:
            $ref: '#/definitions/ErrorMessageDto'
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - apiKey: []
      deprecated: false
definitions:
  PersonResponse:
    type: object
    properties:
      person_uid:
        type: string
        format: uuid
      return_hash:
        type: string
      phone_number:
        type: string
      firstname:
        type: string
      surname:
        type: string
      address_home:
        $ref: '#/definitions/Address'
      email:
        type: string
      health_status:
        $ref: '#/definitions/CodebookItemDto'
      health_status_last_change:
        type: string
  PersonRequest:
    type: object
    properties:
      firstname:
        type: string
      surname:
        type: string
      address_home:
        $ref: '#/definitions/Address'
      email:
        type: string
  SimtompsRequest:
    type: object
    properties:
      symtomps_since:
        type: string
        format: date
      high_temperature_duration:
        type: integer
      dry_cough_duration:
        type: integer
      headache:
        type: boolean
  ExposureRequest:
    type: object
    properties:
      infected_in_contact:
        type: string
        enum:
          - 'Y'
          - 'N'
          - '?'
      infected_in_contact_date:
        type: string
        format: date
      infected_phone_numbers:
        type: string
      visited_risk_area:
        $ref: '#/definitions/CodebookItemDto'
  TestingPlaceRequest:
    type: object
    properties:
      preferred_health_check_location:
        $ref: '#/definitions/CodebookItemDto'
      able_to_drive:
        type: boolean
  HealthCheckRequest:
    type: object
    properties:
      symtomps_since:
        type: string
        format: date
      high_temperature_duration:
        type: integer
      dry_cough_duration:
        type: integer
      headache:
        type: boolean
      infected_in_contact:
        type: string
        enum:
          - 'Y'
          - 'N'
          - '?'
      infected_in_contact_date:
        type: string
        format: date
      infected_phone_numbers:
        type: string
      visited_risk_area:
        $ref: '#/definitions/CodebookItemDto'
      preferred_health_check_location:
        $ref: '#/definitions/CodebookItemDto'
      able_to_drive:
        type: boolean
  HealthCheckResponse:
    type: object
    properties:
      time_slot:
        type: string
      address:
        type: string
      health_check_code:
        type: string
  TreatmentRequest:
    type: object
    properties:
      place:
        $ref: '#/definitions/CodebookItemDto'
      address:
        $ref: '#/definitions/Address'
      health_check_date:
        type: string
        format: date
      health_check_location:
        $ref: '#/definitions/CodebookItemDto'
      expected_infection_date:
        type: string
        format: date
      in_contact_phone_numbers:
        type: string
  Address:
    type: object
    properties:
      street:
        type: string
      street_number_descriptive:
        type: string
      street_number_evidence:
        type: string
      city:
        type: string
      zip_code:
        type: string
  CodebookItemDto:
    type: object
    properties:
      code:
        type: string
      default:
        type: boolean
      order:
        type: integer
        format: int32
      text:
        type: string
    title: CodebookItemDto
  VerifyCodeResponseDto:
    type: object
    properties:
      password:
        type: string
        description: New password
    title: VerifyCodeResponseDto
    description: Sms code verification response
  ErrorMessageDto:
    type: object
    properties:
      additionalMessages:
        type: object
      errorCode:
        type: string
        enum:
          - SMS_CODE_GEN_ERROR
          - UNAUTHORIZED
          - UNKNOWN
      message:
        type: string
    title: ErrorMessageDto

  TestingPlaceInstuctionsDto:
    type: object
    properties:
      address:
        $ref: '#/definitions/Address'
      openingHours:
        type: string